+++ pymtl3/datatypes/bits_import.py
@@ -13,7 +13,7 @@ Date   : Aug 23, 2018
 """
 import os
 
-from pymtl3.utils import custom_exec
+from pymtl3.third_party.pypy import custom_exec
 
 # This __new__ approach has better performance
 # bits_template = """
