+++ pymtl3/passes/sim/PrepareSimPass.py
@@ -15,7 +15,7 @@ from pymtl3.dsl.Connectable import Const, Interface, MethodPort, Signal
 from pymtl3.dsl.NamedObject import NamedObject
 from pymtl3.passes.BasePass import BasePass, PassMetadata
 from pymtl3.passes.errors import PassOrderError
-from pymtl3.utils import custom_exec
+from pymtl3.third_party.pypy import custom_exec
 
 from .SimpleTickPass import SimpleTickPass
